pyproject.toml
[project]
name = "hive-industrial-sentinel"
version = "0.1.0"
description = "Governed AI agents for industrial smart-grid monitoring demos."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "openai>=1.76",
  "pandas>=2.2",
  "pypdf>=4.2",
  "pyyaml>=6.0",
  "streamlit>=1.34",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
